refactor: favorite 패키지를 subway 하위 경로로 이동
