+++ src/main/java/nextstep/subway/favorite/application/FavoriteService.java
@@ -1,14 +1,14 @@
-package nextstep.favorite.application;
+package nextstep.subway.favorite.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
-import nextstep.favorite.application.dto.FavoriteRequest;
-import nextstep.favorite.application.dto.FavoriteResponse;
-import nextstep.favorite.domain.Favorite;
-import nextstep.favorite.domain.FavoriteRepository;
-import nextstep.favorite.exception.DuplicateFavoriteException;
-import nextstep.favorite.exception.UnauthorizedFavoriteAccessException;
+import nextstep.subway.favorite.application.dto.FavoriteRequest;
+import nextstep.subway.favorite.application.dto.FavoriteResponse;
+import nextstep.subway.favorite.domain.Favorite;
+import nextstep.subway.favorite.domain.FavoriteRepository;
+import nextstep.subway.favorite.exception.DuplicateFavoriteException;
+import nextstep.subway.favorite.exception.UnauthorizedFavoriteAccessException;
 import nextstep.member.domain.LoginMember;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
