+++ src/main/java/nextstep/subway/favorite/application/dto/FavoriteRequest.java
@@ -1,4 +1,4 @@
-package nextstep.favorite.application.dto;
+package nextstep.subway.favorite.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
