+++ src/main/java/nextstep/subway/favorite/domain/Favorite.java
@@ -1,4 +1,4 @@
-package nextstep.favorite.domain;
+package nextstep.subway.favorite.domain;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
