+++ src/main/java/nextstep/subway/favorite/web/FavoriteController.java
@@ -1,11 +1,11 @@
-package nextstep.favorite.web;
+package nextstep.subway.favorite.web;
 
 import java.net.URI;
 import java.util.List;
 import lombok.AllArgsConstructor;
-import nextstep.favorite.application.FavoriteService;
-import nextstep.favorite.application.dto.FavoriteRequest;
-import nextstep.favorite.application.dto.FavoriteResponse;
+import nextstep.subway.favorite.application.FavoriteService;
+import nextstep.subway.favorite.application.dto.FavoriteRequest;
+import nextstep.subway.favorite.application.dto.FavoriteResponse;
 import nextstep.member.domain.LoginMember;
 import nextstep.member.ui.AuthenticationPrincipal;
 import org.springframework.http.ResponseEntity;
