+++ src/test/java/nextstep/subway/unit/FavoriteServiceMockTest.java
@@ -6,18 +6,17 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import nextstep.favorite.application.FavoriteService;
-import nextstep.favorite.application.dto.FavoriteRequest;
-import nextstep.favorite.application.dto.FavoriteResponse;
-import nextstep.favorite.domain.Favorite;
-import nextstep.favorite.domain.FavoriteRepository;
-import nextstep.favorite.exception.DuplicateFavoriteException;
-import nextstep.favorite.exception.FavoriteNotFoundException;
-import nextstep.favorite.exception.UnauthorizedFavoriteAccessException;
+import nextstep.subway.favorite.application.FavoriteService;
+import nextstep.subway.favorite.application.dto.FavoriteRequest;
+import nextstep.subway.favorite.application.dto.FavoriteResponse;
+import nextstep.subway.favorite.domain.Favorite;
+import nextstep.subway.favorite.domain.FavoriteRepository;
+import nextstep.subway.favorite.exception.DuplicateFavoriteException;
+import nextstep.subway.favorite.exception.FavoriteNotFoundException;
+import nextstep.subway.favorite.exception.UnauthorizedFavoriteAccessException;
 import nextstep.member.domain.LoginMember;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
